Test and Zip changes
Line endings fixed
+ testAssign no longer opens vs, just builds in command line
+ zipAssign no longer copies and deletes SuperCommon.h

+++ testAssign.sh
@@ -2,7 +2,8 @@
 
 if [[ -z $1 ]]; then
   echo "testAssign -- unzips a zipped assignment to temp dir 'ASSIGN_TEST' and builds and runs it"
-  echo "              deletes temp dir on compleation"
+  echo "              deletes temp dir on compleation."
+  echo "           -- Tests to make sure project builds without environment var CODEUTILS set"
   echo "Usage: testAssign [ziped filename]"
   exit 1
 fi
@@ -12,7 +13,11 @@ if ! [[ -f $1 ]]; then
   exit 1
 fi
 
-rm -r ./ASSIGN_TEST
+if [[ -d "ASSIGN_TEST" ]]; then
+  read -rsp $'About to delete folder ASSIGN_TEST hit any button to continue...\n'
+  rm -r ./ASSIGN_TEST
+fi
+
 mkdir ./ASSIGN_TEST
 cp $1 ./ASSIGN_TEST
 cd ./ASSIGN_TEST
@@ -23,8 +28,8 @@ rm $1
 export CODEUTILS=
 
 for i in *.sln; do
-  echo Building $i
-  devenv $i /Runexit Release
+  #echo "Building $i in Release"
+  devenv.com $i /build Release
 done
 
 read -rsp $'Press enter to continue...\n'
